fix(examples): make cases 113/114/117 self-contained (binary-only)
abicheck's header-dump path runs castxml in C mode (-x c -std=gnu11), which
cannot parse the C++ constructs these cases rely on — [[gnu::abi_tag]],
char8_t / <cstddef>, and [[no_unique_address]] — so the integration and
validate-examples jobs failed at the dump step even though the libraries
compiled fine.

Inline each header's declarations directly into v1.cpp/v2.cpp, drop the
#include and the .h files, and make app.cpp headerless. With no header the
snapshot is taken from the compiled library's DWARF (built -std=c++20 via
CMake), which already carries the correct C++ types — the same binary-only
pattern used for the dual-ABI example. CMakeLists drop the V*_HEADERS args;
docs/examples regenerated.

Verified locally with GCC 13 + castxml (same as CI): case112/113/114/116/117
PASS, case115 SKIP (_BitInt unsupported on GCC 13).
