+++ docs/examples/case113_abi_tag_changed.md
@@ -44,8 +44,6 @@ duplicate noise.
 - `CMakeLists.txt`
 - `app.cpp`
 - `v1.cpp`
-- `v1.h`
 - `v2.cpp`
-- `v2.h`
 
 _See also: [Examples overview](index.md) · [All BREAKING cases](by-verdict/breaking.md) · [Category: Breaking](by-category/breaking.md)._
