+++ docs/examples/case117_no_unique_address.md
@@ -45,8 +45,6 @@ this (and not, say, C++20 modules) is in scope for the snapshot pipeline.
 - `CMakeLists.txt`
 - `app.cpp`
 - `v1.cpp`
-- `v1.h`
 - `v2.cpp`
-- `v2.h`
 
 _See also: [Examples overview](index.md) · [All BREAKING cases](by-verdict/breaking.md) · [Category: Breaking](by-category/breaking.md)._
