+++ examples/case113_abi_tag_changed/CMakeLists.txt
@@ -1,8 +1,6 @@
 abicheck_add_case(case113_abi_tag_changed
     V1_SOURCES v1.cpp
     V2_SOURCES v2.cpp
-    V1_HEADERS v1.h
-    V2_HEADERS v2.h
     APP_SOURCES app.cpp
     PLATFORMS linux macos windows
 )
