+++ examples/case113_abi_tag_changed/v2.cpp
@@ -1,3 +1,12 @@
-#include "v2.h"
+// v2: the abi_tag on get_id() was removed, so its mangled name changes and the
+// old symbol disappears. Self-contained (no header).
 
-int get_id() { return 42; }
+int get_id();   // abi_tag removed
+
+struct Widget {
+    int value;
+};
+
+int get_id() { return 7; }
+
+int widget_value(const Widget *w) { return w->value; }
