--- examples/case113_abi_tag_changed/v2.h
@@ -1,4 +1,0 @@
-#pragma once
-// v2: the same function loses the ABI tag. The demangled declaration is
-// identical, but the mangled symbol is now _Z6get_idv — a distinct symbol.
-int get_id();
