+++ examples/case117_no_unique_address/CMakeLists.txt
@@ -1,8 +1,6 @@
 abicheck_add_case(case117_no_unique_address
     V1_SOURCES v1.cpp
     V2_SOURCES v2.cpp
-    V1_HEADERS v1.h
-    V2_HEADERS v2.h
     APP_SOURCES app.cpp
     PLATFORMS linux macos windows
 )
