+++ examples/case117_no_unique_address/v2.cpp
@@ -1,3 +1,11 @@
-#include "v2.h"
+// v2: EmptyPolicy stored with [[no_unique_address]] — overlaps with the
+//     next member, shrinking Widget. Self-contained (no header).
 
-int widget_value(const Widget *w) { return w->value; }
+struct EmptyPolicy {};
+
+struct Widget {
+    [[no_unique_address]] EmptyPolicy policy;  // overlaps next member
+    long value;
+};
+
+int widget_value(const Widget *w) { return (int)w->value; }
